+++ include/nigiri/routing/query.h
@@ -40,7 +40,7 @@ struct query {
   unsigned min_connection_count_{0U};
   bool extend_interval_earlier_{false};
   bool extend_interval_later_{false};
-  uint8_t profile_{0};
+  uint16_t profile_{0};
 };
 
 }  // namespace nigiri::routing
